+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Change Log
 ==========
 
 * Added `Client` method for the bulk-update properties API endpoint.
+* Added `Property` flag `use_bulk_update` and class method `update_values` to support bulk-update of property values while still using the `value` attribute.
 
 3.3.1 (8JAN19)
 --------------